# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Leveling Bot API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BotStatus {
  online: boolean;
  guildCount: number;
  userCount: number;
  /** Bot latency in ms */
  ping: number;
  /**
     * Uptime in seconds
     * @nullable
     */
  uptime?: number | null;
}

export interface DiscordUser {
  id: string;
  username: string;
  /** @nullable */
  discriminator?: string | null;
  /** @nullable */
  avatar: string | null;
  /** @nullable */
  globalName?: string | null;
}

export interface Guild {
  id: string;
  name: string;
  /** @nullable */
  icon?: string | null;
  botPresent: boolean;
  /** @nullable */
  memberCount?: number | null;
}

export interface ApiError {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}