# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Leveling Bot API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscordUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    DiscordUser:
      type: object
      required: [id, username, avatar]
      properties:
        id:
          type: string
        username:
          type: string
        discriminator:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        globalName:
          type: ["string", "null"]
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
...