# lib/api-zod/src/generated/types/discordUser.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Leveling Bot API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DiscordUser {
  id: string;
  username: string;
  /** @nullable */
  discriminator?: string | null;
  /** @nullable */
  avatar: string | null;
  /** @nullable */
  globalName?: string | null;
}